--- internal/services/ui/.gitignore
@@ -1,1 +1,0 @@
-static/
